upd: Update all entities
