+++ src/main/java/com/zett/hcaredemo/entity/MedicalRecord.java
@@ -31,6 +31,17 @@ public class MedicalRecord {
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
+    @PrePersist
+    protected void onCreate() {
+        createdAt = LocalDateTime.now();
+        updatedAt = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    protected void onUpdate() {
+        updatedAt = LocalDateTime.now();
+    }
+
     @OneToOne
     @JoinColumn(name = "prescription_id")
     private Prescription prescription;
